Setup integration test
