+++ .github/workflows/ci.yml
@@ -12,11 +12,6 @@ jobs:
     runs-on: ubuntu-latest
     env:
       GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
-    services:
-      httpbin:
-        image: kennethreitz/httpbin:latest
-        ports:
-          - 8080:80
     steps:
       - name: Check out source code
         uses: actions/checkout@v3
@@ -31,8 +26,6 @@ jobs:
 
       - name: Run tests
         run: make ci
-        env:
-          HTTPBIN_END_POINT: http://localhost:8080/
 
       - name: Run octocov
         uses: k1LoW/octocov-action@v0
